parses_sites_service/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
cssselect==1.2.0
fake-useragent==1.5.1
idna==3.8
importlib_metadata==8.5.0
lxml==5.3.0
lxml_html_clean==0.2.2
parse==1.20.2
pika==1.3.2
pyee==11.1.1
PyMySQL==1.1.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dotenv==1.0.1
requests==2.32.3
requests-html==0.10.0
soupsieve==2.6
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.20
w3lib==2.2.1
websockets==10.4
zipp==3.20.1
